+++ src/hooks/useDocumentTitle.hook.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function useDocumentTitle (title, fallbackTitle) {
+  React.useEffect(() => {
+    document.title = title;
+    return () => {
+      document.title = fallbackTitle;
+    }
+  }, [title]);
+  return;
+}
+
+export default useDocumentTitle;
